refactor: 기존 전체 내용을 dynamic csr로 받아왔던 것 변경
전체 내용을 csr로 받아오게 되면 페이지 로딩속도가 저하되는 문제가 발생, 그래서 페이지에 보여지는 주소입력 창만 csr방식사용
